+++ .github/workflows/tssdk-ci.yml
@@ -7,6 +7,7 @@ on:
       - packages/sdk-node/**
       - packages/extension-sdk/**
       - packages/extension-sdk-react/**
+      - packages/extension-utils/**
 
   push:
     branches:
@@ -17,6 +18,7 @@ on:
       - packages/sdk-node/**
       - packages/extension-sdk/**
       - packages/extension-sdk-react/**
+      - packages/extension-utils/**
 
   workflow_dispatch:
