+++ packages/api-explorer/package.json
@@ -32,6 +32,7 @@
     "deploy-dev-portal": "rm -f ../../../developer-portal/static/apix/dist/*.js* && cp -R ./dist/*.js* ../../../developer-portal/static/apix/dist"
   },
   "devDependencies": {
+    "@looker/extension-utils": "^0.1.0",
     "@looker/components-test-utils": "^1.5.5",
     "@looker/sdk-codegen-scripts": "^21.1.2",
     "@looker/sdk-node": "^21.18.1",
