+++ packages/api-explorer/src/test-utils/index.ts
@@ -26,4 +26,3 @@
 export * from './lodes'
 export * from './router'
 export * from './redux'
-export { registerTestExtAdaptor } from './extensionAdaptor'
