+++ packages/api-explorer/src/utils/index.ts
@@ -33,4 +33,3 @@ export {
 } from './path'
 export { getLoded } from './lodeUtils'
 export { useWindowSize } from './useWindowSize'
-export * from './extensionAdaptor'
